+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Documentation) Add docs on setting timezone for containers
 - (Bugfix) Ensure that client cache is initialized before using it
 - (Feature) (DBServer Maintenance) Agency adjustments
+- (Logging) Internal client trace
 
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long